refactor(learning): keep C1 R0 boundary narrow
