--- docs/evidence/learning_kernel_r0_dogfood_preflight_v1.json
@@ -1,1 +1,0 @@
-{"candidate_count":3,"candidates":[{"episode_id":"d0.live_agent.credential_boundary_repair","reason_code":"blocked_missing_independent_verdict","record_hash":"sha256:e8d3e17c6de3493bdf4c81a1268886fd19300c75138d39aa80b4b498e64ae652","verification_state":"repair_resolved"},{"episode_id":"d0.live_agent.retry_limit_repair","reason_code":"blocked_missing_independent_verdict","record_hash":"sha256:42503fb674b46f6a7a1a72f635d6f6c949b801edbda2af2e9456f8cc87ba0f04","verification_state":"repair_resolved"},{"episode_id":"d0.live_agent.freeze_reuse","reason_code":"blocked_missing_independent_verdict","record_hash":"sha256:950c386447bd707d720956a8db08cd885376baeb22c401f323e4762233baa17e","verification_state":"locally_verified"}],"development_recommendation":"flat","improvement_claim":false,"projection_hash":"sha256:b4620128d675bc1ba5a0305c5dac266d9936fd8add4f9c8ecbf361ac33f87005","runtime_recommendation":"flat","schema_version":"compass.learning_kernel.dogfood_preflight.v1","source_artifact_hash":"sha256:3d42b762f8ff68d53bc7b57bdfe0fe7c0e01998914908fd0d7f5dda382b2d6f5","source_bundle_hash":"sha256:b83c01032c3680732ac11f1f6bda77f6bb7ba1640d05082072ca4105dbf4060c","stage_a_admitted_count":0,"stage_a_state":"blocked_missing_independent_verdict","synthesized_authority":{"capsule":false,"impact":false,"reward":false,"selector":false,"utility":false}}
