--- docs/evidence/learning_kernel_r0_mechanism_summary_v1.json
@@ -1,1 +1,0 @@
-{"causal":{"candidate_delta":0.25,"permutation_deltas":[0.0,0.0],"permutation_p95":0.0},"decision":{"candidate_state":"candidate_only","failed_metric":null,"improvement_claim":false,"observed_value":null,"reason_code":"all_learning_gates_green","runtime_recommendation":"flat","threshold_value":null},"efficiency":{"latency_p50_ms":9.0,"latency_p95_ms":10.0,"total_estimated_cost_usd":0.0,"total_input_tokens":4330,"total_output_tokens":672},"expected_journal_hash":"sha256:93bca7d70f6659227521adadaa74afcbe2a44820636aa71f8fdc91a1988deb64","fixture_id":"lkr0_fixture_r0","fixture_manifest_hash":"sha256:aede3cca28c92ff64b25f576a3e05a74f9c1124fd91931f8a910da85942b19f0","forgetting":{"recovery_rate":0.0,"regret":0.04166666666666674},"improvement_claim":false,"interventions":["no_memory","raw","distilled","shuffled","stale","contradictory","poisoned"],"journal_hash":"sha256:93bca7d70f6659227521adadaa74afcbe2a44820636aa71f8fdc91a1988deb64","matrix_runs":336,"runtime_recommendation":"flat","safety":{"admitted_poisoned_view_ids":[],"protected_deltas":{"protected":0.0}},"schema_version":"compass.learning_kernel.summary.v1","selectors":["flat","semantic","distilled","contextual_utility","current_poi","governed"],"summary_hash":"sha256:0a7369a3119e725d28497a6763e49320b4e6f2c0650a3fb2b7ea6f9c76275fd7"}
