--- docs/evidence/learning_kernel_r0_protocol_v1.json
@@ -1,99 +1,0 @@
-{
-  "schema_version": "compass.learning_kernel.protocol.v1",
-  "protocol_hash": "sha256:382af6209caa2c47c44b73ddfb440ad9d979789b05c387d2849e0ba497cc8cea",
-  "code_commit": "72ad7f81d42311b8a89619e3571310185372354c",
-  "fixture_manifest_hash": "sha256:aede3cca28c92ff64b25f576a3e05a74f9c1124fd91931f8a910da85942b19f0",
-  "mechanism_summary_file_hash": "sha256:de0a9333c6303d306e92b57320d27d8378956cc8797790cff593e18a4edc6758",
-  "dependency_lock": {
-    "path": "pyproject.toml",
-    "sha256": "sha256:a6d3c50880d6a60129142f4910afe908886bf3fbe49c94f4ba61eaec27773109"
-  },
-  "python_version": "3.13.14",
-  "input_contract": {
-    "source": "frozen mechanically verified fixtures and Ed25519-authenticated verdict bindings",
-    "provider_probe_is_evidence": false,
-    "caller_supplied_trust_registry_allowed": false
-  },
-  "local_gate": {
-    "status": "passed",
-    "runtime_recommendation": "flat",
-    "improvement_claim": false,
-    "learning_kernel": {
-      "passed": 82,
-      "deselected_pending_protocol": 1,
-      "command": "python -m pytest tests/learning_kernel_r0 -q -k not\\ protocol_binds"
-    },
-    "gep": {
-      "passed": 355,
-      "command": "python -m pytest tests/gep -q"
-    },
-    "poi_non_live": {
-      "passed": 211,
-      "deselected_live": 535,
-      "command": "python -m pytest tests/poi_gate2 -q -k not\\ live_agent"
-    },
-    "poi_live_split": {
-      "covered": 535,
-      "passed": 535,
-      "strategy": "file and query-class splits to avoid monolithic orchestration timeout",
-      "components": {
-        "cli_and_evidence": 53,
-        "execution": 59,
-        "freeze": 61,
-        "freeze_source_eol": 4,
-        "git": 10,
-        "harness": 80,
-        "l0_fixture": 6,
-        "probe_readback_bundle_runner": 15,
-        "schema": 140,
-        "stream": 15,
-        "task_verifier": 13,
-        "verifiers": 79
-      },
-      "transient_observation": {
-        "first_status": "12 passed and 1 temporary repository-root resolution failure under concurrent load",
-        "isolated_replay": "the failed process-tree timeout case passed in 16.79 seconds",
-        "hidden": false
-      }
-    },
-    "monolithic_poi_probe": {
-      "status": "timeout",
-      "interpretation": "not treated as passing evidence; complete split coverage is recorded above"
-    },
-    "ruff": {
-      "status": "passed",
-      "scope": "Learning Kernel, PoI and GEP source plus tests"
-    },
-    "diff_check": "passed"
-  },
-  "mechanism_gate": {
-    "matrix_runs": 336,
-    "candidate_delta": 0.25,
-    "permutation_p95": 0.0,
-    "protected_delta": 0.0,
-    "protected_regression_floor": -0.0005,
-    "admitted_poison_count": 0,
-    "candidate_state": "candidate_only"
-  },
-  "independent_review": {
-    "initial_high_findings": 3,
-    "initial_medium_findings": 3,
-    "high_findings": 0,
-    "medium_findings": 0,
-    "final_scope": "verdict authentication, admission, replay integrity, fixture execution boundary, forgetting and policy eligibility",
-    "final_focused_test": "7 passed"
-  },
-  "external_gate": {
-    "status": "not_started",
-    "reason": "internal causal delta must remain reproducible before external baseline or SOTA comparison",
-    "platform_v2_vectors": "pending",
-    "super_agent_adapter": "not_implemented"
-  },
-  "permissions": {
-    "platform_or_fde_write": false,
-    "super_agent_runtime_connection": false,
-    "deployment": false,
-    "push_or_merge": false,
-    "model_weight_training": false
-  }
-}
